Promote LogRecordBuilder#setBody(AnyValue) to stable API
