+++ api/all/src/main/java/io/opentelemetry/api/common/AnyValueBytes.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.api.incubator.logs;
+package io.opentelemetry.api.common;
 
 import java.nio.ByteBuffer;
 import java.util.Arrays;
