+++ api/incubator/src/main/java/io/opentelemetry/api/incubator/events/DefaultEventLogger.java
@@ -5,8 +5,8 @@
 
 package io.opentelemetry.api.incubator.events;
 
+import io.opentelemetry.api.common.AnyValue;
 import io.opentelemetry.api.common.Attributes;
-import io.opentelemetry.api.incubator.logs.AnyValue;
 import io.opentelemetry.api.logs.Severity;
 import io.opentelemetry.context.Context;
 import java.time.Instant;
