+++ api/incubator/src/test/java/io/opentelemetry/api/incubator/logs/AnyValueTest.java
@@ -9,6 +9,9 @@ import static io.opentelemetry.sdk.testing.assertj.OpenTelemetryAssertions.asser
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.params.provider.Arguments.arguments;
 
+import io.opentelemetry.api.common.AnyValue;
+import io.opentelemetry.api.common.AnyValueType;
+import io.opentelemetry.api.common.KeyAnyValue;
 import java.nio.ByteBuffer;
 import java.nio.ReadOnlyBufferException;
 import java.nio.charset.StandardCharsets;
