+++ exporters/otlp/common/src/main/java/io/opentelemetry/exporter/internal/otlp/AnyValueMarshaler.java
@@ -5,8 +5,8 @@
 
 package io.opentelemetry.exporter.internal.otlp;
 
-import io.opentelemetry.api.incubator.logs.AnyValue;
-import io.opentelemetry.api.incubator.logs.KeyAnyValue;
+import io.opentelemetry.api.common.AnyValue;
+import io.opentelemetry.api.common.KeyAnyValue;
 import io.opentelemetry.exporter.internal.marshal.MarshalerWithSize;
 import java.nio.ByteBuffer;
 import java.util.List;
